test(events): split tests into multiple modules
